+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 4.1.0
 
+* added visitors, sessions, entry/exit rate, and average time on page to entry/exit page statistics when filtering for an event
 * upgraded to Go version 1.19
 * switched from `atomic.LoadInt32` to `atomic.Bool`
-* added visitors, sessions, and entry/exit rate to entry/exit page statistics when filtering for an event
 * updated dependencies
 
 ## 4.0.0